chore(rakki): Split components
